requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.4
pydantic[email]==2.6.4
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
radon==6.0.1
javalang==0.13.0
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
scikit-learn==1.4.1.post1
nltk==3.8.1
transformers==4.38.2
torch==2.2.1
GitPython==3.1.42
codecarbon==2.3.5
python-dotenv==1.0.1
